extract useMask from useLoadingPage
